Social Login

+++ src/main/java/com/akiponkuma/blog/presentation/controller/PostController.java
@@ -1,12 +1,15 @@
 package com.akiponkuma.blog.presentation.controller;
 
 import java.time.LocalDateTime;
-import java.util.Date;
 import java.util.List;
 
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.social.connect.Connection;
+import org.springframework.social.connect.ConnectionData;
+import org.springframework.social.connect.ConnectionRepository;
+import org.springframework.social.facebook.api.Facebook;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -17,11 +20,19 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import com.akiponkuma.blog.domain.model.entity.Post;
 import com.akiponkuma.blog.infrastructure.dao.PostDao;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
 @Controller
 @RequestMapping("/post")
 public class PostController {
 	@Autowired
 	private PostDao postDao;
+	
+	@Autowired
+	private ConnectionRepository connectionRepository;
 	
 	@RequestMapping(value="/write", method=RequestMethod.GET)
 	public String form(Post post){
@@ -31,13 +42,17 @@ public class PostController {
 	
 	@RequestMapping(value="/write", method=RequestMethod.POST)
 	public String write(@Valid Post post, BindingResult bindingResult){
+		User user = getConnect();		
 		if(bindingResult.hasErrors()){
 			return "form";
 		}		
 		post.setRegDate(LocalDateTime.now());
+		post.setUserId(user.getProviderUserId());
+		post.setName(user.getDisplayName());
 		return "redirect:/post/" + postDao.save(post).getId();
 	}
 	
+
 	@RequestMapping("/list")
 	public String list(Model model){
 		List<Post> postList = postDao.findAll();
@@ -54,7 +69,11 @@ public class PostController {
 	
 	@RequestMapping("/{id}/delete")
 	public String delete(@PathVariable int id){
-		postDao.delete(id);
+		User user = getConnect();
+		Post post = postDao.findOne(id);
+		if(post.getUserId().equals(user.getProviderUserId())){
+			postDao.delete(id);			
+		}		
 		return "redirect:/post/list";
 	}
 	
@@ -67,9 +86,36 @@ public class PostController {
 	
 	@RequestMapping(value="/{id}/edit", method=RequestMethod.POST)
 	public String edit(@Valid Post post, BindingResult bindingResult){
+		User user = getConnect();
 		if(bindingResult.hasErrors()){
-			return "form";			
+			return "form";		
 		}
-		return "redirect:/post/" + postDao.save(post).getId();
+		
+		Post oldPost = postDao.findOne(post.getId());
+		if(oldPost.getUserId().equals(user.getProviderUserId())){
+			oldPost.setTitle(post.getTitle());
+			oldPost.setSubtitle(post.getSubtitle());
+			oldPost.setContent(post.getContent());
+			return "redirect:/post/" + postDao.save(oldPost).getId();
+		}
+		return "form";
+	}
+	
+	private User getConnect() {
+		Connection<Facebook> connection = connectionRepository.findPrimaryConnection(Facebook.class);
+		if(connection == null){
+			return null;			
+		}
+		ConnectionData data = connection.createData();
+		return new User(data.getProviderUserId(), data.getDisplayName());
+	}
+	
+	@Getter
+	@Setter
+	@NoArgsConstructor
+	@AllArgsConstructor
+	public static class User {
+		String providerUserId;
+		String displayName;
 	}
 }
